Fix alignment issues with memcpy etc.  Gcc is using a memcpy optimised for aligned structures based on the types of the pointers passed.  This generates alignment traps on ARM (and probably other CPUs) because the pointers are not aligned.
This is a partially overlooked issue when moving from packed structures in
Squashfs 3.x and older (which are not guaranteed to be aligned) to un-packed
structures in Squashfs 4.0 which are assumed to be aligned.  Only partially
overlooked because you used to be able to assume if the function prototype
pointers were declared as char * (like memcpy), then passed pointers wouldn't
be treated as aligned.