Change memcpy to memmove.
If an inode is large (>16K) it will cause the src and dest pointers
to overlap in the data cache memcpy, potentially leading to block_list
corruption/truncation.

This code dates back to Mksquashfs v1.0 (2002) where inodes were not
very large, inodes can now be considerably larger and this latent bug
should have been fixed.

Given the nature of the bug it surprising it hasn't caused any issues
until now.