action: introduce macros for the test fn boilerplate
Introduce macros for the generic test function, and for the
test function which compares a file attribute for being equal, less than
or greater than a value.

This allows multiple tests to be implemented without constantly
reimplementing the boilerplate.

Also reimplement the name_fn and the filesize_fn tests using
these macros.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>