actions: add empty action
Exclude the directory if it is empty, and the expression returns true.

This action is added to deal with situations where excluding results in
empty directories (including situations where a directory may contain
other directories but whuch recursively are ultimately empty).

These directories once the contents have been excluded are often completely
useless and should be excluded themselves.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>