mksquashfs: add subpath to dir_info structure
Rather than pathname which includes the path specified on the
command line, add a subpath which is rooted at the root of the
squashfs filesystem being created.

i.e. given mksquashfs a/b file.sqsh

with file "c/d/e" inside directory "b"

the pathname of "d" will be "a/b/c/d"

this adds a subpath which will be "c/d"

This is necessary for the new pathname test for actions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>