mksquashfs: Use a linked list to store directory entries rather than an array
When scanning and storing the source directories use a
linked list to store the directory entries rather than an array.

This change is to more efficiently handle the empty action and the
soon to be implemented move action.  The empty action removes
directory entries and the move action may move directory entries
from one directory to another.  These actions are rather messy to
implement with arrays requiring any holes resulting from deleting/moving
to be filled by shuffling array entries down to fill the hole.
Linked lists allow entries to be deleted/moved much easier, which
is not only a performance improvement but results in more
aesthetically pleasing code.

Arrays were previously used only because the qsort function requires
arrays.  The move to linked lists has improved the code and this is
an additional unexpected bonus, which is in addition to the previously
mentioned reasons to move to linked lists.

Qsort is stil used in this commit, and so code has been added to
convert the linked-list into an array before calling qsort, and code
has also been added to convert the sorted array into a linked-list.
This is messy, but only temporary, to ensure this commit doesn't
break mksquashfs.  The next commit will remove qsort and will instead
add a merge sort which can efficiently sort linked lists
( O(n log n) which is the same complexity as qsort).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>