mksquashfs: fix assignment of source_name
Normally create_dir_entry will be called with source_name == NULL
if the file has not been renamed.

If the file has been renamed, create_dir_entry will be called
with name holding the new name, and source_name holding the
original_name.

This commit fixes the callers which in the case where the file
has not been renamed, call create_dir_entry with source_name ==
name, leading to source_name being != NULL when the file has not
been renamed.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>