mksquashfs: fix -info to use subpathname()
Now we have subpathname() that represents the pathname of the file
from the point of view of the output Squashfs filesystem (rather
than pathname(), which represents the pathname in the source
filesystem), use that in preference to pathname().

In the future with the move action, which can rename and move files
around, this is necessary because the output should reflect the
results of any move actions.

Additionally, irrespective of the new move action, subpathname()
was always a better choice than pathname(), but until recently it
didn't exist, and so couldn't be used.

Also fix the numerous -info outputs which incorrectly just output
the name of the file rather than the full pathname.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>