pseudo: fix handling of modify pseudo files
Matthieu CASTET reported a bug with modify pseudo files where if
a directory was modified, it would be subsequently treated as
a non-directory for further pseudo modify files.

His provided patch was unfortunately only a partial fix.

This is a full fix for the bug.

Thanks to Matthieu CASTET for reporting the bug.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>