mksquashfs: Add -progress to force progress bar display when using -info
Add -progress option to enable progress bar display when
using -info option.

Originally when the progress bar was introduced in mksquashfs 3.2,
it was always displayed, even with the -info option.  To disable
the progress bar you had to specify -no-progress.  This although
logically correct, seemed counter-intuitive as in the majority of
cases you didn't want the progress bar displayed with -info.
So, in mksquashfs 3.4 -info was changed to disable the
progress bar by default.

Unfortunately, a by-product of this change was to make it
impossible to have both the progress bar displayed and -info
output in the rare cases where it is desirable (sometimes you
want to both see where mksquashfs has got to in the source
directories being Squashed, and where that is/mksquashfs is
in the overall Squashfs filesystem creation).

So finally, after over 4 years, add back in the ability to
display both the progress bar and -info output.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>