actions: on malloc failures bail out using BAD_ERROR()
On malloc failures we have little option but to bail out because
these failures are unrecoverable.  The only other option is to ignore
these failures, and to treat the actions or tests where they occur
as having failed.  However, if these actions or tests would have done
something if the malloc failure didn't occur, we end up with a
filesystem which is wrong - not actually corrupted but one which
is unexpected, which is worse than explicitly bailing out.

In practice with the default over-commit behaviour of the Linux kernel,
mallocs rarely fail, and the first thing a process knows about out of
memory situations is when it is killed by the OOM (out of memory)
killer.

It is perhaps worth mentioning this is why the recovery file was added
on appending, because it was discovered on out of memory situations,
mksquashfs never got the chance to restore the filesystem being appended to.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>