mksquashfs: check for trailing junk after k/K/m/M
Tighten up parsing for -b block number.

Originally Mksquashfs deliberately didn't check for trailing junk
after the k/K/m/M as this allowed things such as 1Mbyte to be
parsed.  However, this means rubbish such as 1mjjjsjdlj is parsed
as OK.

Given the whole point of the current exercise is to tighten up
{un/mk}squasfs' detection of bogus user input after CVEs were raised
against it, it seems silly to keep this feature.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>