pseudo: fix potential stack overflow in get_component()
This fix is similar to the fix for CVE-2012-4024 raised against
unsquashfs, except mksquashfs was not covered by CVE-2012-4024

Fix potential stack overflow in get_component() where an individual
pathname component in a pseudo file (specified on the command line
or in a pseudo file) could exceed the 1024 byte sized targname
allocated on the stack.

Fix by dynamically allocating targname rather than storing it as
a fixed size on the stack.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>