sort: tighten up sort file parsing
When reviewing read_sort_file() for buffer overflow, I noticed
it did not deal with the case where the sort file entry had
a missing priority.

Also deal with the unlikely case that fscanf() returns 0.

Plus make the existing case where the priority is outside the range
-32767:32768 a fatal error rather than skipping the entry, this
IMHO is a more appropriate response, and also improve the error
statement too.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>