read_fs: fix small memory leaks in read_filesystem()
Fix some memory leaks in read_filesystem()...  Mostly
these are in the error path (where mksquashfs will shortly abort).

There is one (id_table) which is not in the error path, and so
this missing free may have leaked a couple of Kbytes when using
Mksquashfs to append to a filesystem.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>