actions: fix lexeme to string mapping
The insertion of the '\t' whitespace lexeme in the token_table[]
broke the lexeme value to string mapping...   The mapping relies
on the fact each lexeme value corresponds to the index in
the token_table, i.e. the lexeme value is the same as the
lexeme's index in the token_table.

Move all the whitespace lexemes to the end of the table,
to ensure additional whitespace lexemes can be added in
future without breaking the mapping.  Change the lexeme
values so that they correspond to the new order.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>