unsquashfs: add code to only print "not superuser" error message once
When running as non-superuser if unsquashfs encounters a non
user.xxx xattr, it will print an error message as it cannot
extract these xattrs whilst running as a normal user process.

Unfortunately, for many filesystems, every file can be inadertantly
tagged with a security.xxx xattr, with the direct result trying to
unsquash these results in screenfulls of the same error message!

The previous commit added a -user-xattr option to suppress these error
messages, but this commit goes further, and adds code to
suppress the error message after it printed once, and adds advice
on how to prevent this one error message from being printed in the future!

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>