mksquashfs: flag lseek error in write_destination as probable out of space
If the main thread hits "lseek error" when writing the metadata
in write_destination() it is probably because the
destination file is out of space, this is especially true if the
destination is a block device, but could conceivably happen on some
(non Linux?) filesystems too if they don't implement holes.  This
is because lseeking to the write position could be beyond the end
of the available destination file end, and hence fail before
attempting to write.

Recent correspondence with users shows that "Lseek error" is
meaningless, and so augment the error message with an "Out of
Space" error which is presumbly something they will understand.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>